// app/utility/type.ts
 export type CovidData = {
    dateRep: string;
    day: string;
    month: string;
    year: string;
    cases: number;
    deaths: number;
    countriesAndTerritories: string;
    geoId: string;
    countryterritoryCode: string;
    popData2019: number;
    continentExp: string;
    Cumulative_number_for_14_days_of_COVID_19_cases_per_100000: string;
  };